requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
nltk==3.6.2
scikit-learn==0.24.2
transformers==4.9.2
torch==1.9.0
spacy==3.1.2
pytest==6.2.5
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
gunicorn==20.1.0